+++ src/main/java/com/example/app/common/config/OwnerCheckFilter.java
@@ -1,41 +1,33 @@
 package com.example.app.common.config;
 
 import com.example.app.common.util.AuthUtil;
-import com.example.app.userinfo.auth.AuthService;
-import com.example.app.userinfo.auth.SessionStorage;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
+import com.example.app.auth.service.AuthService;
+import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
 @RequiredArgsConstructor
-public class OwnerCheckFilter extends OncePerRequestFilter {
+public class OwnerCheckFilter implements Filter {
 
     private final AuthService authService;
-    private final SessionStorage sessionStorage;
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        HttpServletRequest request = (HttpServletRequest) servletRequest;
+        HttpServletResponse response = (HttpServletResponse) servletResponse;
         try {
             Long resourceId = Long.valueOf(AuthUtil.extractResourceId(request.getRequestURI()));
-
             Long currentUserId = authService.getCurrentUserId(request);
-
             boolean isOwner = authService.isOwner(resourceId, currentUserId);
             if (!isOwner) {
                 response.setStatus(HttpServletResponse.SC_FORBIDDEN);
                 response.getWriter().write("접근 권한 없음");
                 return;
             }
-
             filterChain.doFilter(request, response);
-
         } catch (NumberFormatException e) {
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             response.getWriter().write("잘못된 리소스 ID 형식");
@@ -43,7 +35,7 @@ public class OwnerCheckFilter extends OncePerRequestFilter {
             response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
             response.getWriter().write(e.getMessage());
         }
-    }
 
+    }
 
 }
