+++ src/main/java/com/example/app/userinfo/entity/UserInfo.java
@@ -1,8 +1,9 @@
 package com.example.app.userinfo.entity;
 
 
-import com.example.app.common.entity.BaseEntity;
+import com.example.app.common.model.BaseEntity;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -20,11 +21,13 @@ public class UserInfo extends BaseEntity {
     private String userName;
 
     @Column(nullable = false, unique = true)
-//    @Email(message = "이메일 형식에 맞게 작성")
+    @Email(message = "이메일 형식에 맞게 작성")
     private String email;
 
     @Column(nullable = false)
     private String password;
+
+    // casecade 연습해보기
 
     public UserInfo(String userName, String email, String password){
         this.userName = userName;
